requirements.txt
nba_api
pandas
numpy
scikit-learn
xgboost
lightgbm
matplotlib
seaborn
joblib
python-dotenv
groq # For Groq API
streamlit
streamlit-option-menu
psycopg2-binary
SQLAlchemy